Move component AgencyCard from 'components/cards' to 'components/index/CardsReference'
